pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kayak-travel-recommender"
version = "0.1.0"
description = "Travel destination recommender based on weather data"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.0",
    "boto3>=1.40.0",
    "pandas>=2.3.0",
    "numpy>=2.0.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "streamlit>=1.53.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "flake8>=7.3.0",
]
scraping = [
    # scrapy-playwright 0.0.45 fixes Scrapy 2.14 compatibility (HTTP11DownloadHandler import path)
    # Requires: Python >=3.10, Scrapy >=2.7, Playwright >=1.40
    "scrapy>=2.14.0",
    "scrapy-playwright>=0.0.45",
    "playwright>=1.40.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-vvv", "-m", "not integration"]
pythonpath = ["src"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
